+++ .env
@@ -1,6 +1,6 @@
-REACT_APP_BACKEND_URL='https://api.consumet.org/'
-#  miruro-api.herokuapp.com/ Self-hosted consumet.org API
-#* https://miruro-api.up.railway.app/ Self-hosted Gogoanime API
-#* http://localhost:32768/ Local entry point base URL
+VITE_BACKEND_URL='http://localhost:3000/'
+#?  miruro-api.herokuapp.com/ Self-hosted api.consumet.org API [MAIN]
+#* https://miruro-api.up.railway.app/ Self-hosted Gogoanime API [SECONDARY]
+#* http://localhost:3000/ Local entry point base URL
 #! https://api.consumet.org/ Rate limited
-#! https://sakamoto-api.vercel.app/ Decrepated/Discontinued API
+#! https://sakamoto-api.vercel.app/ Decrepated/Discontinued API - Currently used for Video Streaming in Production
