+++ .env.example
@@ -1,1 +1,8 @@
-REACT_APP_BACKEND_URL=Self-hosted api from https://api.consumet.org/ 
+# Rename this file to .env for environment variables
+
+VITE_BACKEND_URL='http://localhost:3000/'
+#?  miruro-api.herokuapp.com/ Self-hosted api.consumet.org API [MAIN]
+#* https://miruro-api.up.railway.app/ Self-hosted Gogoanime API [SECONDARY]
+#* http://localhost:3000/ Local entry point base URL
+#! https://api.consumet.org/ Rate limited
+#! https://sakamoto-api.vercel.app/ Decrepated/Discontinued API - Currently used for Video Streaming in Production
