+++ docs/index.html
@@ -1,74 +1,72 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/x-icon" href="/assets/img/favicon.ico">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" type="image/x-icon" href="/assets/img/favicon.ico" />
     <!--=============== BOXICONS ===============-->
-    <link href='https://unpkg.com/boxicons@2.1.2/css/boxicons.min.css' rel='stylesheet'>
+    <link
+      href="https://unpkg.com/boxicons@2.1.2/css/boxicons.min.css"
+      rel="stylesheet"
+    />
 
     <!--=============== CSS ===============-->
-    <link rel="stylesheet" href="assets/css/styles.css">
+    <link rel="stylesheet" href="assets/css/styles.css" />
 
     <title>Miruro</title>
-</head>
+  </head>
 
-<body>
+  <body>
     <!--==================== HEADER ====================-->
     <header class="header">
-        <nav class="nav container">
-            <a href="https://miruro.vercel.app" class="nav__logo">
-                見るろ
-            </a>
+      <nav class="nav container">
+        <a href="https://miruro.vercel.app" class="nav__logo"> 見るろ </a>
 
-            <div class="nav__menu" id="nav-menu">
-                <ul class="nav__list">
-                    <li class="nav__item">
-                        <a href="pages/Privacy.html" class="nav__link">Privacy</a>
-                    </li>
-                    <li class="nav__item">
-                        <a href="pages/ToS.html" class="nav__link">ToS</a>
-                    </li>
-                    <li class="nav__item">
-                        <a href="https://miruro.vercel.app" class="nav__link">Website</a>
-                    </li>
-                </ul>
+        <div class="nav__menu" id="nav-menu">
+          <ul class="nav__list">
+            <li class="nav__item">
+              <a href="pages/Privacy.html" class="nav__link">Privacy</a>
+            </li>
+            <li class="nav__item">
+              <a href="pages/ToS.html" class="nav__link">ToS</a>
+            </li>
+            <li class="nav__item">
+              <a href="https://miruro.vercel.app" class="nav__link">Website</a>
+            </li>
+          </ul>
 
-                <div class="nav__close" id="nav-close">
-                    <i class='bx bx-x'></i>
-                </div>
-            </div>
+          <div class="nav__close" id="nav-close">
+            <i class="bx bx-x"></i>
+          </div>
+        </div>
 
-            <!-- Toggle button -->
-            <div class="nav__toggle" id="nav-toggle">
-                <i class='bx bx-grid-alt'></i>
-            </div>
-        </nav>
+        <!-- Toggle button -->
+        <div class="nav__toggle" id="nav-toggle">
+          <i class="bx bx-grid-alt"></i>
+        </div>
+      </nav>
     </header>
 
     <!--==================== MAIN ====================-->
     <main class="main">
-        <!--==================== HOME ====================-->
-        <section class="home">
-            <div class="home__container container">
-                <div class="home__data">
-                    <span class="home__subtitle">Welcome To</span>
-                    <h1 class="home__title">MIRURO DOCS</h1>
-                    <p class="home__description">
-                        Miruro Documentation<br>
-                    </p>
-                    <a href="https://miruro.vercel.app" class="home__button">
-                        Miruro
-                    </a>
-                </div>
+      <!--==================== HOME ====================-->
+      <section class="home">
+        <div class="home__container container">
+          <div class="home__data">
+            <span class="home__subtitle">Welcome To</span>
+            <h1 class="home__title">MIRURO DOCS</h1>
+            <p class="home__description">Miruro Documentation<br /></p>
+            <a href="https://miruro.vercel.app" class="home__button">
+              Miruro
+            </a>
+          </div>
 
-                <div class="home__img">
-                    <img src="assets/img/logo512.png" alt="">
-                    <div class="home__shadow"></div>
-                </div>
-            </div>
-        </section>
+          <div class="home__img">
+            <img src="assets/img/logo512.png" alt="" />
+            <div class="home__shadow"></div>
+          </div>
+        </div>
+      </section>
     </main>
 
     <!--=============== SCROLLREVEAL ===============-->
@@ -76,6 +74,5 @@
 
     <!--=============== MAIN JS ===============-->
     <script src="assets/js/main.js"></script>
-</body>
-
+  </body>
 </html>
