+++ src/index.jsx
@@ -10,14 +10,5 @@ root.render(
   </React.StrictMode>
 );
 
-//* Function to disable right click on website
-/* window.addEventListener(
-  'contextmenu',
-  function (e) {
-    e.preventDefault();
-  },
-  false
-); */
-
 // register the service worker
 serviceWorkerRegistration.register();
