+++ src/pages/Top100Anime.jsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState, useRef } from "react";
 import { Link } from "react-router-dom";
-import styled from "styled-components";
+import styled, { css } from "styled-components"; // Import 'css' from 'styled-components'
 import SearchResultsSkeleton from "../components/Skeletons/SearchResultsSkeleton";
 import axios from "axios";
 
-function RecentEpisodesAnime() {
-
+function Top100Anime() {
   // State variables
   const [animeDetails, setAnimeDetails] = useState([]);
   const [loading, setLoading] = useState(true);
@@ -69,17 +68,16 @@ function RecentEpisodesAnime() {
   // Function to fetch anime data
   async function getAnime(page) {
     try {
-      const response = await axios.get(
-        `${process.env.REACT_APP_BACKEND_URL}anime/gogoanime/recent-episodes?page=${page}&perPage=30`
-      );
+      const response = await axios.get(`https://api.jikan.moe/v4/top/anime`, {
+        params: {
+          /* type: "tv", // You can change the type as needed */
+          page,
+          limit: 25,
+        },
+      });
 
-      if (response.data && response.data.results) {
-        setAnimeDetails((prevData) => [
-          ...prevData,
-          ...response.data.results.filter((item) =>
-            prevData.every((prevItem) => prevItem.id !== item.id)
-          ),
-        ]);
+      if (response.data && response.data.data) {
+        setAnimeDetails((prevData) => [...prevData, ...response.data.data]);
         setCurrentPage(page);
       } else {
         console.error("Invalid response structure:", response.data);
@@ -102,23 +100,21 @@ function RecentEpisodesAnime() {
   // JSX for rendering
   return (
     <div>
-      {loading && <SearchResultsSkeleton name="Recent Episodes Anime" />}
+      {loading && <SearchResultsSkeleton name="Top 100 Anime" />}
       <Parent>
         <Heading>
-          <span>Recent Episodes Anime</span> Results
+          <span>Top 100 Anime</span> Results
         </Heading>
         <CardWrapper>
           {animeDetails.map((item, i) => (
             <Wrapper to={`/search/${item.title}`} key={i}>
-              <img className="card-img" src={item.image} alt="" />
-              <p>
-                {item.title.romaji ||
-                  item.title.english ||
-                  item.title.native ||
-                  item.title.userPreferred ||
-                  item.title}
-              </p>
-              <p>{"Episode: " + item.episodeNumber || 'Unknown Type'}</p>
+              <img
+                className="card-img"
+                src={item.images.jpg.image_url}
+                alt=""
+              />
+              <p>{item.title}</p>
+              <p>{item.type || "Unknown Type"}</p>
             </Wrapper>
           ))}
         </CardWrapper>
@@ -133,15 +129,16 @@ function RecentEpisodesAnime() {
 
 // Styled components
 
-// ... styles for Parent component
+// Styles for the Parent component
 const Parent = styled.div`
   margin: 2rem 5rem 2rem 5rem;
+
   @media screen and (max-width: 600px) {
-    margin: 1rem;
+    margin: 1rem; // Adjusted margin for smaller screens
   }
 `;
 
-// ... styles for CardWrapper component
+// Styles for the CardWrapper component
 const CardWrapper = styled.div`
   display: grid;
   grid-template-columns: repeat(auto-fill, 160px);
@@ -168,34 +165,44 @@ const CardWrapper = styled.div`
   }
 `;
 
+// Common styles for image elements (used by Wrapper and Links components)
+const CommonImageStyles = css`
+  width: 160px;
+  height: 235px;
+  border-radius: 0.4rem;
+  object-fit: cover;
+
+  @media screen and (max-width: 600px) {
+    width: 120px;
+    height: 180px;
+    border-radius: 0.3rem;
+  }
+
+  @media screen and (max-width: 400px) {
+    width: 110px;
+    height: 170px;
+  }
+
+  @media screen and (max-width: 380px) {
+    width: 100px;
+    height: 160px;
+  }
+`;
+
+// Styles for the Wrapper component
 const Wrapper = styled(Link)`
   text-decoration: none;
   img {
-    width: 160px;
-    height: 235px;
-    border-radius: 0.4rem;
-    object-fit: cover;
-    @media screen and (max-width: 600px) {
-      width: 120px;
-      height: 180px;
-      border-radius: 0.3rem;
-    }
-    @media screen and (max-width: 400px) {
-      width: 110px;
-      height: 170px;
-    }
-    @media screen and (max-width: 380px) {
-      width: 100px;
-      height: 160px;
-    }
+    ${CommonImageStyles}// Reuse common image styles
   }
 
   p {
     color: #ffffff;
     font-size: 1rem;
-    font-family: 'Gilroy-Medium', sans-serif;
+    font-family: "Gilroy-Medium", sans-serif;
     text-decoration: none;
     max-width: 160px;
+
     @media screen and (max-width: 380px) {
       width: 100px;
       font-size: 0.9rem;
@@ -203,27 +210,11 @@ const Wrapper = styled(Link)`
   }
 `;
 
-// ... styles for Links component
+// Styles for the Links component (similar to Wrapper)
 const Links = styled(Link)`
   text-decoration: none;
   img {
-    width: 160px;
-    height: 235px;
-    border-radius: 0.4rem;
-    object-fit: cover;
-    @media screen and (max-width: 600px) {
-      width: 120px;
-      height: 180px;
-      border-radius: 0.3rem;
-    }
-    @media screen and (max-width: 400px) {
-      width: 110px;
-      height: 170px;
-    }
-    @media screen and (max-width: 380px) {
-      width: 100px;
-      height: 160px;
-    }
+    ${CommonImageStyles}// Reuse common image styles
   }
 
   p {
@@ -232,6 +223,7 @@ const Links = styled(Link)`
     font-family: "Gilroy-Medium", sans-serif;
     text-decoration: none;
     max-width: 160px;
+
     @media screen and (max-width: 380px) {
       width: 100px;
       font-size: 0.9rem;
@@ -239,12 +231,13 @@ const Links = styled(Link)`
   }
 `;
 
-// ... styles for Heading component
+// Styles for the Heading component
 const Heading = styled.p`
   font-size: 1.8rem;
   color: #fff;
   font-family: "Gilroy-Light", sans-serif;
   margin-bottom: 2rem;
+
   span {
     font-family: "Gilroy-Bold", sans-serif;
   }
@@ -255,4 +248,4 @@ const Heading = styled.p`
   }
 `;
 
-export default RecentEpisodesAnime;
+export default Top100Anime;
